Move all optics to the same modules, update tests
 * define encapsulation for internally-used optics
